Improve types and move to own files
