+++ src/workflows/pdf/PdfWorkflowManager.ts
@@ -33,14 +33,17 @@ type FinalizePdfHandler<
   } & FactoryMeta<TPdfFactoryMeta>
 ) => number /* pageNumber */;
 
+type FactoryGenerator<TFactoryData, TPdfFactoryMeta extends PdfFactoryMeta> =
+  | Constructable<PdfFactory<TFactoryData, TPdfFactoryMeta>>
+  | PdfFactoryPicker<TFactoryData, TPdfFactoryMeta>;
+
 export default class PdfWorkflowManager<
   TFactoryData,
-  TPdfFactoryMeta extends PdfFactoryMeta,
-  TFactory extends PdfFactory<TFactoryData, TPdfFactoryMeta>
+  TPdfFactoryMeta extends PdfFactoryMeta
 > extends WorkflowManager {
   protected data: TFactoryData[];
   protected entityIds: number[] = [];
-  protected factories: TFactory[] = [];
+  protected factories: PdfFactory<TFactoryData, TPdfFactoryMeta>[] = [];
   protected factoriesMeta: Map<number, FactoryMeta<TPdfFactoryMeta>> =
     new Map();
 
@@ -60,9 +63,7 @@ export default class PdfWorkflowManager<
   }
 
   constructor(
-    factory:
-      | Constructable<TFactory>
-      | PdfFactoryPicker<TFactoryData, TPdfFactoryMeta, TFactory>,
+    factory: FactoryGenerator<TFactoryData, TPdfFactoryMeta>,
     data: TFactoryData[],
     extractEntityId: (data: TFactoryData) => number
   ) {
